Ref: Outil de purge des fichiers vides.
git-svn-id: svn://svn.code.sf.net/p/mediboard/code/trunk@13467 a21d6be2-e512-0410-a060-91f5a2946631

+++ modules/dPfiles/ajax_repair_files.php
@@ -8,30 +8,57 @@
  * @license GNU General Public License, see http://www.gnu.org/licenses/gpl.html 
  */
 
-$nb_files = round(CValue::get("nb_files", 0));
-$file = new CFile();
+$nb_files     = round(CValue::get("nb_files", 0));
+$date_debut   = CValue::get("date_debut");
+$date_fin     = CValue::get("date_fin");
+$purge = CValue::get("purge", 0);
 
-$file->file_size = 0;
+$file = new CFile;
 
-$nb_files_size_zero = $file->countMatchingList();
-
-if ($nb_files == 0) {
-  CAppUI::stepAjax("Nombre de fichiers avec une taille de 0 octets : " . $nb_files_size_zero);  
-}
-else {
-  $files = $file->loadMatchingList(null, $nb_files);
-  
-  if (count($files) == 0 ) {
-    CAppUI::stepAjax("Aucun fichier à traiter");
-  }
-  else {
-    foreach ($files as &$_file) {
-      if (!file_exists($_file->_file_path) || filesize($_file->_file_path) === 0) {
-        CAppUI::stepAjax("File id : " . $_file->_id . " - non existant ou vide - Suppression :" . $_file->delete());
+if ($date_debut && $date_fin) {
+  $where = array();
+  $where["file_date"] =  "BETWEEN '".mbDateTime($date_debut)."' AND '".mbDateTime($date_fin)."'";
+  $count_files = $file->countList($where);
+  $files = $file->loadList($where, null, $nb_files);
+  if ($purge) {
+    $count = 0;
+    foreach ($files as $_file) {
+      if ($msg = $_file->purge()) {
+        CAppUI::stepAjax("File id: " . $_file->_id . " - " . $_file->purge());
       }
       else {
-        $_file->file_size = filesize($_file->_file_path);
-        CAppUI::stepAjax("File id : " . $_file->_id . " - mise à jour de la taille ({$_file->file_size} octets)- Update :" . $_file->store());
+        $count++;
+      }
+    }
+    CAppUI::stepAjax("$count fichiers supprimés");
+  }
+  else {
+    CAppUI::stepAjax("$count_files fichiers à traiter");
+  }
+}
+else {
+  $file->file_size = 0;
+  $nb_files_size_zero = $file->countMatchingList();
+  if ($nb_files == 0) {
+    CAppUI::stepAjax("Nombre de fichiers avec une taille de 0 octets : " . $nb_files_size_zero);  
+  }
+  else {
+    $where = array();
+    $where["file_size"] = " = '0'";
+    $files = $file->loadList($where);
+    
+    if (count($files) == 0 ) {
+      CAppUI::stepAjax("Aucun fichier à traiter");
+    }
+    else {
+      foreach ($files as &$_file) {
+        if (!file_exists($_file->_file_path) || filesize($_file->_file_path) === 0) {
+          CAppUI::stepAjax("File id : " . $_file->_id . " - non existant ou vide - Suppression :" . $_file->delete());
+        }
+        else {
+          $_file->file_size = filesize($_file->_file_path);
+          CAppUI::stepAjax("File id : " . $_file->_id . " - mise à jour de la taille ({$_file->file_size} octets)- Update :" . $_file->store());
+        }
       }
     }
   }
